# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/api/users": {
      "post": {
        "operationId": "UsersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Post User",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenUserDto"
                }
              }
            }
          }
        },
        "tags": [
          "users"
        ]
      }
    },
    "/api/batches": {
      "post": {
        "operationId": "BatchesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateBatchDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "BatchesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/batches/{id}": {
      "get": {
        "operationId": "BatchesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "BatchesController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateBatchDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "BatchesController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/companies": {
      "post": {
        "operationId": "CompaniesController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCompanyDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "CompaniesController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/companies/{id}": {
      "get": {
        "operationId": "CompaniesController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/invitations": {
      "post": {
        "operationId": "InvitationsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        }
      },
      "get": {
        "operationId": "InvitationsController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
    "/api/invitations/{id}": {
      "get": {
        "operationId": "InvitationsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "patch": {
        "operationId": "InvitationsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateInvitationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        }
      },
      "delete": {
        "operationId": "InvitationsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    }
  },
  "info": {
    "title": "Composer API",
    "description": "Composer API description",
    "version": "0.0",
    "contact": {}
  },
  "tags": [
    {
      "name": "composer",
      "description": ""
    }
  ],
  "servers": [],
  "components": {
    "schemas": {
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "TokenUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "id": {
            "type": "number"
          },
          "accessToken": {
            "type": "string"
          }
        },
        "required": [
          "email",
          "id",
          "accessToken"
        ]
      },
      "CreateBatchDto": {
        "type": "object",
        "properties": {
          "lotNumber": {
            "type": "string"
          },
          "parentLotNumber": {
            "type": "string"
          },
          "leadContent": {
            "type": "number"
          },
          "mercuryContent": {
            "type": "number"
          },
          "cadmiumContent": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "unit": {
            "type": "object"
          }
        },
        "required": [
          "lotNumber",
          "parentLotNumber",
          "leadContent",
          "mercuryContent",
          "cadmiumContent",
          "quantity",
          "unit"
        ]
      },
      "UpdateBatchDto": {
        "type": "object",
        "properties": {
          "lotNumber": {
            "type": "string"
          },
          "parentLotNumber": {
            "type": "string"
          },
          "leadContent": {
            "type": "number"
          },
          "mercuryContent": {
            "type": "number"
          },
          "cadmiumContent": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          },
          "unit": {
            "type": "object"
          }
        }
      },
      "CreateCompanyDto": {
        "type": "object",
        "properties": {}
      },
      "CreateInvitationDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateInvitationDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}